@@ -2,13 +2,10 @@
     <nav id="side-bar">
         <ul class="side-bar-list">
 
-            <li v-for="list in titleList">
-                <p class="list-title-primary">{{list.primaryTitle}}</p>
-
-                <p :class="{'active': arg == $index}"
-                    v-for="subList in list.secondTitle">
-
-                    <a v-link="list.basePath + subList.link" @click="setActive($index)">{{subList.title}}</a>
+            <li v-for="list in routesList.list">
+                <p class="list-title-primary">{{list.name}}</p>
+                <p>
+                    <router-link :to="'/' + routesList.parent + '/' + list.path">{{list.path}}</router-link>
                 </p>
             </li>
 
@@ -17,49 +14,16 @@
 </template>
 
 <script>
-
 export default {
     data() {
         return {
-            arg: '',
-            titleList: [
-                {
-                    primaryTitle: 'config-base',
-                    basePath: '/webpack-vue/',
-                    secondTitle: [
-                        {
-                            title: 'overview',
-                            link: 'base-overview'
-                        }
-                    ]
-                },
-                {
-                    primaryTitle: 'config-dev',
-                    basePath: '/webpack-vue/',
-                    secondTitle: [
-                        {
-                            title: 'overview',
-                            link: 'dev-overview'
-                        }
-                    ]
-                },
-                {
-                    primaryTitle: 'config-build',
-                    basePath: '/webpack-vue/',
-                    secondTitle: [
-                        {
-                            title: 'overview',
-                            link: 'build-overview'
-                        }
-                    ]
-                }
-            ]
-        }
+        };
     },
 
-    methods: {
-        setActive(index) {
-            this.arg = index;
+    props: {
+        routesList: {
+            type: Object,
+            required: true
         }
     }
 }
@@ -67,30 +31,23 @@ export default {
 
 <style lang="scss">
 @import '../scss/_global.scss';
-
 #side-bar {
     @include border(right);
-
     a {
         display: block;
         padding-left: 28px;
     }
-
     p {
         margin: 8px 0;
     }
-
     .side-bar-list {
         padding-left: 10px;
         padding-right: 10px;
         @extend %list-style-none;
-
     }
-
     .list-title-primary {
         padding-left: 15px;
         @extend %pos-r;
-
         &:before {
             width: 3px;
             height: 100%;
@@ -100,11 +57,8 @@ export default {
             @extend %pos-t-l;
         }
     }
-
     .active {
         color: $color_e;
     }
 }
-
-
 </style>
